Handle new enum type change in interfaces.json
When using the API call `repository.create` the required `vcs` argument is no
longer an <int> but a <string>.  This meant the mapping used by
`parse_interfaces` (enum: int) would throw the following error::

  ValueError: Wrong argument type: vcs is not a <type 'int'>

This has been updated in the mapping of datatypes, and a small
doctest added.

Also:

I have made the text of the error slightly clearer (in my view), as I was
initially confused as to whether I was provding an int or should be an int.

I have altered the module docstring to use >> not >>> to stop upsetting
doctest.

I have put too many changes into one commit :-)